+++ src/main/java/com/flowiee/pms/repository/product/ProductComboRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductComboRepository extends JpaRepository<ProductCombo, Integer> {
+public interface ProductComboRepository extends JpaRepository<ProductCombo, Long> {
 }
